Fix nested any/all evaluation
The current evaluator pops everything on the stack in order to evaluate `any` or `all` queries. That breaks in nested situations like:

```
all(any(unix, target_arch="x86"), not(any(target_os="android", target_os="emscripten")))
```

Fix this by tracking the number of predicates in the `Any` and `All` variants.

This unfortunately breaks semver because the definition of `Func` changes.
`Func` doesn't seem to be exported anywhere so maybe in the upcoming release
it could be made private?

+++ src/error.rs
@@ -16,7 +16,7 @@ pub struct ParseError<'a> {
 #[derive(Debug, PartialEq)]
 pub enum Reason {
     /// not() takes exactly 1 predicate, unlike all() and any()
-    InvalidNot(u8),
+    InvalidNot(usize),
     /// The characters are not valid in an SDPX license expression
     InvalidCharacters,
     /// An opening parens was unmatched with a closing parens
